refactor(backstop-mcp): extract client helpers into utils.py, route credential checks through BackstopClient
Splits path/header/deserialization helpers out of client.py so verify_credential can reuse
the real BackstopClient (and its retry/error handling) instead of hand-rolling a request, and
widens metric_route to label every id segment instead of dropping the full path.
